+++ tools/src/commands/GenerateDocsAPIData.ts
@@ -125,6 +125,7 @@ const PACKAGES_MAPPING: Record<string, CommandAdditionalParams> = {
   'expo-router': ['exports.ts'],
   'expo-router-ui': ['ui/index.ts', 'expo-router'],
   'expo-router-native-tabs': ['native-tabs/index.ts', 'expo-router'],
+  'expo-router-split-view': ['split-view/index.ts', 'expo-router'],
   'expo-screen-capture': ['ScreenCapture.ts'],
   'expo-screen-orientation': ['ScreenOrientation.ts'],
   'expo-secure-store': ['SecureStore.ts'],